add js

+++ src/main/java/com/bbs/bigmud/bbs/dto/QuestionDTO.java
@@ -7,12 +7,12 @@ import lombok.Data;
 @Data
 public class QuestionDTO {
 
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmt_Create;
     private Long gmt_Modified;
-    private Integer creater;
+    private Long creater;
     private Integer Comment_Count;
     private Integer Like_Count;
     private Integer View_Count;
